Revert "refactor: added EmailConfirmationService and edited tests"
This reverts commit eb3af7656c01ad110cbdc865a13d0488dfc8ba59.
