+++ tests/unit/services/userService.test.ts
@@ -103,6 +103,29 @@ describe('UserService', () => {
 		});
 	});
 
+	describe('verifyUserEmail', () => {
+		const user = users[0];
+		beforeEach(async () => {
+			await getRepository(User).save(user);
+		});
+
+		test('Successfully verifies a user', async () => {
+			const verifiedUser = await userService.verifyUserEmail(user.id);
+			expect(verifiedUser.accountStatus).toStrictEqual(AccountStatus.Verified);
+		});
+
+		test('Fails when trying to verify an already-verified user', async () => {
+			const verifiedUser = await userService.verifyUserEmail(user.id);
+			expect(verifiedUser.accountStatus).toStrictEqual(AccountStatus.Verified);
+			await expect(userService.verifyUserEmail(user.id)).rejects.toMatchObject({ httpCode: HttpCode.BadRequest });
+		});
+
+		test('Fails with invalid user ID', async () => {
+			await expect(userService.verifyUserEmail('')).rejects.toMatchObject({ httpCode: HttpCode.NotFound });
+			await expect(userService.verifyUserEmail(`${user.id}123`)).rejects.toMatchObject({ httpCode: HttpCode.NotFound });
+		});
+	});
+
 	describe('authenticate', () => {
 		const user = users[0];
